requirements.txt
Django==1.7.1
Pillow==2.7.0
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-ckeditor==5.0.2
django-storages==1.1.8
django-suit==0.2.12
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
